refactor(components/atom/image): add import file extensions
