+++ components/atom/image/demo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "atom-image-demo",
+  "name": "@s-ui/react-atom-image-demo",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
