+++ components/atom/image/src/settings.js
@@ -1,0 +1,7 @@
+export const BASE_CLASS = 'sui-AtomImage'
+export const BASE_CLASS_FIGURE = `${BASE_CLASS}-figure`
+export const CLASS_PLACEHOLDER = `${BASE_CLASS_FIGURE}--placeholder`
+export const CLASS_SKELETON = `${BASE_CLASS_FIGURE}--skeleton`
+export const CLASS_IMAGE = `${BASE_CLASS}-image`
+export const CLASS_SPINNER = `${BASE_CLASS}-spinner`
+export const CLASS_ERROR = `${BASE_CLASS}-error`
